pyproject.toml
[project]
name = "wave-mcp"
version = "1.0.0"
description = "MCP server for Wave accounting — 6 tools covering accounts, customers, vendors, invoices, products, sales taxes, and synthesized reports"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["server.py", "client.py", "auth_flow.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
